backend: Move domain code from core->api
