+++ packages/backend-api/bin/osmod.js
@@ -20,7 +20,6 @@ limitations under the License.
  *
  * Provides commandline access to various features, including
  *  - managing users
- *  - inventing comments for test purposes
  *  - managing comments
  *
  *  For a full list of available commands, run `osmod.js --help`
@@ -34,10 +33,14 @@ const yargs = require('yargs');
 yargs
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'users', 'create')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'users', 'get_token')))
-  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'tests', 'youtube')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'rescore')))
   .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'send_to_scorer')))
-  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'generate')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'calculate_text_size')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'recalculate_text_sizes')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'recalculate_top_scores')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'flag')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'comments', 'delete')))
+  .command(require(path.join(__dirname, '..', 'dist', 'commands', 'denormalize')))
   .demand(1)
   .usage('Usage: $0')
   .help()
