+++ packages/backend-api/package.json
@@ -26,6 +26,7 @@
     "@types/kue": "0.11.1",
     "@types/lodash": "4.14.136",
     "@types/mysql": "2.15.5",
+    "@types/opentype.js": "0.7.2",
     "@types/passport": "1.0.0",
     "@types/passport-google-oauth2": "0.1.2",
     "@types/passport-jwt": "3.0.1",
@@ -51,6 +52,7 @@
     "lodash": "4.17.14",
     "moment": "2.21.0",
     "mysql": "2.15.0",
+    "opentype.js": "1.1.0",
     "passport": "0.4.0",
     "passport-google-oauth20": "2.0.0",
     "passport-jwt": "4.0.0",
