+++ packages/backend-api/src/commands/articles/delete.ts
@@ -16,39 +16,40 @@ limitations under the License.
 
 import * as yargs from 'yargs';
 
-import { denormalizeCommentCountsForArticle } from '../../domain/articles';
-import { logger } from '../../logger';
+import { logger } from '@conversationai/moderator-backend-core';
 import {
   Article,
-  Comment,
-} from '../../models';
+  Category,
+} from '@conversationai/moderator-backend-core';
 
-export const command = 'comments:delete';
-export const describe = 'Delete all comments from the database.';
+import { denormalizeCommentCountsForCategory } from '../../domain';
+
+export const command = 'articles:delete';
+export const describe = 'Delete all articles from the database.';
 
 export function builder(args: yargs.Argv) {
   return args
-    .usage('Usage: node $0 comments:delete');
+    .usage('Usage: node $0 articles:delete');
 }
 
 export async function handler() {
-  logger.info(`Deleting comments`);
+  logger.info(`Deleting articles`);
 
   try {
-    await Comment.destroy({where: {}});
-    const articles = await Article.findAll();
-    for (const a of articles) {
-      logger.info('Denormalizing article ' + a.id);
-      denormalizeCommentCountsForArticle(a, false);
+    await Article.destroy({where: {}});
+    const categories = await Category.findAll();
+    for (const c of categories) {
+      logger.info('Denormalizing category ' + c.id);
+      denormalizeCommentCountsForCategory(c);
     }
 
   }
   catch (err) {
-    logger.error('Delete comments error: ', err.name, err.message);
+    logger.error('Delete articles error: ', err.name, err.message);
     logger.error(err.errors);
     process.exit(1);
   }
 
-  logger.info('Comments successfully deleted');
+  logger.info('Articles successfully deleted');
   process.exit(0);
 }
