+++ packages/backend-api/src/test/integration/api/actions.spec.ts
@@ -29,11 +29,11 @@ import {
   ICommentInstance,
   IUserInstance,
 } from '@conversationai/moderator-backend-core';
+
 import {
   denormalizeCommentCountsForArticle,
   denormalizeCountsForComment,
-} from '@conversationai/moderator-backend-core';
-
+} from '../../../domain';
 import {
   expect,
   makeArticle,
