+++ packages/backend-core/package.json
@@ -5,8 +5,6 @@
   "main": "dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "test": "npm run mocha",
-    "mocha": "../../node_modules/.bin/ts-mocha 'src/test/**/*.spec.ts' --recursive --timeout 200000",
     "lint": "find src -name '*.ts' | xargs ../../node_modules/.bin/tslint -c ../../tslint.json",
     "lint:fix": "find src -name '*.ts' | xargs ../../node_modules/.bin/tslint -c ../../tslint.json --fix",
     "compile": "rm -rf dist && ../../node_modules/.bin/tsc --sourceMap --outDir dist --declaration",
@@ -22,23 +20,15 @@
     "@types/js-yaml": "3.12.1",
     "@types/lodash": "4.14.136",
     "@types/mysql": "2.15.5",
-    "@types/request": "2.47.0",
     "@types/sequelize": "3.4.43",
-    "@types/underscore.string": "0.0.38",
     "@types/yargs": "13.0.2",
     "express-winston": "2.5.1",
-    "find-up": "3.0.0",
-    "googleapis": "32.0.0",
     "joi": "10.5.2",
     "js-yaml": "3.13.1",
     "lodash": "4.17.14",
-    "moment": "2.21.0",
     "mysql": "2.15.0",
-    "opentype.js": "0.6.9",
-    "request": "2.87.0",
     "sequelize": "3.30.4",
     "sequelize-cli": "2.7.0",
-    "underscore.string": "3.3.5",
     "winston": "2.3.1",
     "yargs": "13.3.0"
   }
