+++ packages/backend-core/src/index.ts
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export * from './domain';
 export * from './models';
 export * from './logger';
 export * from './sequelize';
